Use page mode in permitting actions
Before we only used the mode of the namespace. Now the page's mode and
owner are also considered. It's Unix-ish, in that you must have read
access to the namespace and write access to the page if you're modifying
a page.

To read a page, you can have just read on both namespace and page
(there's no execute requirement here).

You need write access to a namespace to create a page, but not to edit
one.